+++ src/features/my-page/pages/artist-profile.tsx
@@ -1,3 +1,20 @@
+import { useState } from 'react';
+
+import { ProfileForm } from '../components/profile-form';
+import { Input } from '../components/ui/input';
+
 export const ArtistProfile = () => {
-  return <div>작가 개인정보</div>;
+  const [instagram, setInstagram] = useState('');
+
+  return (
+    <ProfileForm>
+      <Input
+        id="instagram"
+        label="Instagram Business Account ID"
+        type="text"
+        value={instagram}
+        onChange={(e) => setInstagram(e.target.value)}
+      />
+    </ProfileForm>
+  );
 };
